requirements.txt
Django==1.10.3
django-bootstrap3==7.1.0
django-ckeditor==5.0.3
django-haystack==2.5.1
elasticsearch==5.0.1
